requirements.txt
torch
torchvision 
torchaudio 
marker-pdf
langchain-text-splitters 
langchain_community
langchain-pinecone
